Server: Improve delete endpoint & store revision metadata onto nightly revision
Now all the delete functions return the deleted resource
(release, app, extension).
Moreover the subfolder (revision) in the default release (Nightly) has a 'revision'
metadata that is used for getting it.
This commit correct also Bugs on the delete releases endpoint by allowing the delete
of revision release, that wasn't handled properly before.