Python Client: Create an Application support the choice of the Collection
It's now possible to choose the collection within create the application. It's also
possible to get an existing collection by ID for creating the application inside.

In this case, you must provide the 'coll_id' argument to use all the other
command on these application. By default all the command look for application
which are under the "Applications" collection.