PythonClient: Use tabulate to display results and upgrade/fix client issues
Here a list of different point fixed by this commit:

* Remove the 'DESCRIPTION' field when list resources (extensions, packages...)
* Put 'REVISION' or 'APP REVISION' at the first column when the list of
resources is displayed (as extensions, packages, releases...)
* Add the 'app_revision' info on the confirmation message when a release is
created or deleted
* The 'name' and 'revision' are now required argument of the command `release create`
* Fix the display of extension using the tabulate python feature
* Fix the getting of release_name when displaying the list of extension, this
upgrade the efficiency of the package listing too
* Using the `--all` option when listing extensions, list every single extension
from the application, the `limit` is set to 0.

This commit fix a Server issue too: it now permits to sort from the most recent
extension to the less recent, this bugs occurred when using `release=draft` when
listing all the extension of the draft release.